EY-3274 Flytte feilreg til egen fil
